Refactored a bit
