+++ middleware/middleware/api/project_api.py
@@ -1,11 +1,12 @@
 from flask import request
 
-from ..server import (
-    app,
+from ..server import app
+from ..service import (
     proxy_to_api,
-    requires_auth,
     get_available_projects,
 )
+
+from ..decorator import requires_auth
 
 
 # Site Selection
